# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type MarketAssetKind = typeof MarketAssetKind[keyof typeof MarketAssetKind];


export const MarketAssetKind = {
  crypto: 'crypto',
  rwa: 'rwa',
} as const;

export interface MarketAsset {
  id: number;
  name: string;
  symbol: string;
  kind: MarketAssetKind;
  /** @nullable */
  price: number | null;
  /** @nullable */
  change24h: number | null;
  /** @nullable */
  marketCap: number | null;
  /** @nullable */
  volume24h: number | null;
  /** @nullable */
  rank: number | null;
  color: string;
  /** @nullable */
  imageUrl: string | null;
}

export interface MarketPulse {
  label: string;
  value: number;
  change24h: number;
}

export interface MarketOverview {
  asOf: string;
  totalMarketCap: number;
  marketCapChange24h: number;
  totalVolume24h: number;
  btcDominance: number;
  assets: MarketAsset[];
  pulse: MarketPulse[];
  source: string;
}

export interface MarketExplanationInput {
  /**
     * @minLength 3
     * @maxLength 500
     */
  question: string;
  /**
     * @maxLength 20
     * @nullable
     */
  assetSymbol?: string | null;
}

export interface MarketExplanation {
  answer: string;
  question: string;
  asOf: string;
  source: string;
}

export type LimitParameter = number;

export type GetMarketOverviewParams = {
/**
 * @minimum 3
 * @maximum 12
 */
limit?: LimitParameter;
};

export type GetMarketAssetsParams = {
kind?: GetMarketAssetsKind;
/**
 * @minimum 3
 * @maximum 12
 */
limit?: LimitParameter;
};

export type GetMarketAssetsKind = typeof GetMarketAssetsKind[keyof typeof GetMarketAssetsKind];


export const GetMarketAssetsKind = {
  crypto: 'crypto',
  rwa: 'rwa',
} as const;

export type SearchMarketAssetsParams = {
/**
 * @minLength 2
 * @maxLength 80
 */
q: string;
kind?: SearchMarketAssetsKind;
/**
 * @minimum 3
 * @maximum 12
 */
limit?: LimitParameter;
};

export type SearchMarketAssetsKind = typeof SearchMarketAssetsKind[keyof typeof SearchMarketAssetsKind];


export const SearchMarketAssetsKind = {
  crypto: 'crypto',
  rwa: 'rwa',
} as const;