# lib/api-zod/src/generated/types/marketAsset.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MarketAssetKind } from './marketAssetKind';

export interface MarketAsset {
  id: number;
  name: string;
  symbol: string;
  kind: MarketAssetKind;
  /** @nullable */
  price: number | null;
  /** @nullable */
  change24h: number | null;
  /** @nullable */
  marketCap: number | null;
  /** @nullable */
  volume24h: number | null;
  /** @nullable */
  rank: number | null;
  color: string;
  /** @nullable */
  imageUrl: string | null;
}