# lib/api-zod/src/generated/types/marketExplanationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MarketExplanationInput {
  /**
     * @minLength 3
     * @maxLength 500
     */
  question: string;
  /**
     * @maxLength 20
     * @nullable
     */
  assetSymbol?: string | null;
}