# lib/api-zod/src/generated/types/marketOverview.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MarketAsset } from './marketAsset';
import type { MarketPulse } from './marketPulse';

export interface MarketOverview {
  asOf: Date;
  totalMarketCap: number;
  marketCapChange24h: number;
  totalVolume24h: number;
  btcDominance: number;
  assets: MarketAsset[];
  pulse: MarketPulse[];
  source: string;
}